add validation

+++ auth/internal/api/auth/delete.go
@@ -3,12 +3,20 @@ package auth
 import (
 	desc "auth/pkg/user_v1"
 	"context"
+	"google.golang.org/grpc/codes"
+	"google.golang.org/grpc/status"
 	"google.golang.org/protobuf/types/known/emptypb"
 	"log"
 )
 
 func (s *Implementation) Delete(ctx context.Context, req *desc.DeleteRequest) (*emptypb.Empty, error) {
 	log.Printf("Received user UpdateRequest %+v", req.GetId())
+
+	if req.GetId() == 0 {
+		return nil, status.Errorf(
+			codes.InvalidArgument,
+			"user id is required")
+	}
 
 	if err := s.authService.Delete(ctx, req.GetId()); err != nil {
 		return nil, err
